backend/pyproject.toml
[project]
name = "ai-reel-captioning-backend"
version = "0.1.0"
description = "FastAPI backend for AI Reel Captioning and Intro Zoom Editor"
requires-python = ">=3.11"
dependencies = [
  "alembic>=1.13.2",
  "celery[redis]>=5.4.0",
  "fastapi>=0.115.0",
  "python-multipart>=0.0.12",
  "pydantic-settings>=2.6.1",
  "psycopg[binary]>=3.2.3",
  "sqlalchemy>=2.0.36",
  "uvicorn[standard]>=0.32.0",
  "python-slugify>=8.0.4",
  "httpx>=0.28.1",
  "whisperx>=3.3.1",
  "torch>=2.5,<2.6",
  "torchaudio>=2.5,<2.6",
  "PyJWT>=2.9.0",
  "bcrypt>=4.2.0",
  "email-validator>=2.2.0",
]

[dependency-groups]
dev = [
  "pytest>=8.3.3",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]
